move windower to utils/dictionary
