+++ alphacsc/utils/dictionary.py
@@ -105,6 +105,49 @@ def get_lambda_max(X, D_hat, sample_weights=None):
         ] for D_k in D_hat], axis=(1, 2))[:, None]
 
 
+class NoWindow():
+
+    def window(self, d):
+        return d
+
+    def dewindow(self, d):
+        return d
+
+
+class UVWindower(NoWindow):
+    def __init__(self, n_times_atom, n_channels):
+        self.n_channels = n_channels
+        self.tukey_window = tukey_window(n_times_atom)[None, :]
+
+    def window(self, d):
+        d = d.copy()
+        d[:, self.n_channels:] *= self.tukey_window
+        return d
+
+    def dewindow(self, d):
+        d = d.copy()
+        d[:, self.n_channels:] /= self.tukey_window
+        return d
+
+    def simple_window(self, d):
+        return d * self.tukey_window
+
+    def simple_dewindow(self, d):
+        return d / self.tukey_window
+
+
+class SimpleWindower(NoWindow):
+    def __init__(self, n_times_atom):
+        print(f"initialized {n_times_atom}")
+        self.tukey_window = tukey_window(n_times_atom)[None, None, :]
+
+    def window(self, d):
+        return d * self.tukey_window
+
+    def dewindow(self, d):
+        return d / self.tukey_window
+
+
 def tukey_window(n_times_atom):
     window = signal.tukey(n_times_atom)
     window[0] = 1e-9
